requirements.txt
django==3.2
djangorestframework
django-rest-swagger==2.1.2
python-decouple
psycopg2
django-rest-passwordreset
django-cors-headers